[+ES] Add modules switcher to the header
Note that as part of this we need to work out what modules a user has access
to.

BE VERY CAREFUL HERE! There are two places in the older client_portal apps
where this may come from:

  1. The JSON returned from a call to /profile
  2. The JSON encoded within the JWT access_token

These basically have the same permissions data structured in two different
ways. We're using the JWT ONLY from now on, because calls to /profile are
unnecessary and pointless (given the data is in the token).