Modify the trkref filters to fetch all available ones from profile
The trkrefs filter from Tableau cannot be used to populate the list of trkrefs because it will limit the number of trkrefs to 200 and it will not contain the Retailer name. Instead, we can use the list of trkrefs that are coming back from the /profile request done to client_portal-admin.